requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pytube==10.6.0
typing-extensions==3.7.4.3
